requirements.txt
Pillow==8.4.0
pandas==1.2.3

